+++ packages/client/src/Client/Exception/ConnectionException.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Utopia\Client\Exception;
-
-class ConnectionException extends NetworkException {}