+++ packages/client/src/Client/Exception/DnsException.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Utopia\Client\Exception;
-
-final class DnsException extends NetworkException {}